* Refactoring tool part classes into packages to start adding support for different types of materials
