+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/ToolPartModelType.java
@@ -2,7 +2,7 @@ package com.sigmundgranaas.forgero.client.forgerotool.model;
 
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
 import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPart;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartHead;
+import com.sigmundgranaas.forgero.core.tool.toolpart.head.ToolPartHead;
 
 import java.util.Locale;
