+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/toolpart/handle/HandleModel2D.java
@@ -2,7 +2,7 @@ package com.sigmundgranaas.forgero.client.forgerotool.model.toolpart.handle;
 
 import com.sigmundgranaas.forgero.client.forgerotool.model.ToolPartModelType;
 import com.sigmundgranaas.forgero.client.forgerotool.model.toolpart.Unbaked2DToolPartModel;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartHandle;
+import com.sigmundgranaas.forgero.core.tool.toolpart.handle.ToolPartHandle;
 import net.minecraft.client.render.model.ModelLoader;
 import net.minecraft.client.texture.Sprite;
 import net.minecraft.client.util.SpriteIdentifier;
