+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/toolpart/head/HeadModel2D.java
@@ -2,7 +2,7 @@ package com.sigmundgranaas.forgero.client.forgerotool.model.toolpart.head;
 
 import com.sigmundgranaas.forgero.client.forgerotool.model.ToolPartModelType;
 import com.sigmundgranaas.forgero.client.forgerotool.model.toolpart.Unbaked2DToolPartModel;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartHead;
+import com.sigmundgranaas.forgero.core.tool.toolpart.head.ToolPartHead;
 import net.minecraft.client.render.model.ModelLoader;
 import net.minecraft.client.texture.Sprite;
 import net.minecraft.client.util.SpriteIdentifier;
