+++ src/main/java/com/sigmundgranaas/forgero/core/tool/ForgeroTool.java
@@ -2,8 +2,8 @@ package com.sigmundgranaas.forgero.core.tool;
 
 import com.sigmundgranaas.forgero.core.identifier.tool.ForgeroToolIdentifier;
 import com.sigmundgranaas.forgero.core.material.material.PrimaryMaterial;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartHandle;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartHead;
+import com.sigmundgranaas.forgero.core.tool.toolpart.handle.ToolPartHandle;
+import com.sigmundgranaas.forgero.core.tool.toolpart.head.ToolPartHead;
 import net.minecraft.item.ToolMaterial;
 import org.jetbrains.annotations.NotNull;
