+++ src/main/java/com/sigmundgranaas/forgero/core/tool/ForgeroToolTypes.java
@@ -13,4 +13,8 @@ public enum ForgeroToolTypes {
         }
         return false;
     }
+
+    public String getToolName() {
+        return this.name().toLowerCase();
+    }
 }
