+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/ForgeroToolPartCollectionImpl.java
@@ -1,7 +1,10 @@
 package com.sigmundgranaas.forgero.core.tool.toolpart;
 
 import com.sigmundgranaas.forgero.core.material.MaterialCollection;
+import com.sigmundgranaas.forgero.core.tool.toolpart.binding.ToolPartBinding;
 import com.sigmundgranaas.forgero.core.tool.toolpart.factory.ForgeroToolPartFactory;
+import com.sigmundgranaas.forgero.core.tool.toolpart.handle.ToolPartHandle;
+import com.sigmundgranaas.forgero.core.tool.toolpart.head.ToolPartHead;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
