--- src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/ToolPartBinding.java
@@ -1,4 +1,0 @@
-package com.sigmundgranaas.forgero.core.tool.toolpart;
-
-public interface ToolPartBinding extends ForgeroToolPart {
-}
