--- src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/ToolPartHandle.java
@@ -1,4 +1,0 @@
-package com.sigmundgranaas.forgero.core.tool.toolpart;
-
-public interface ToolPartHandle extends ForgeroToolPart {
-}
