+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/binding/ToolPartBinding.java
@@ -1,0 +1,6 @@
+package com.sigmundgranaas.forgero.core.tool.toolpart.binding;
+
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPart;
+
+public interface ToolPartBinding extends ForgeroToolPart {
+}
