+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/factory/ToolPartBindingBuilder.java
@@ -1,8 +1,8 @@
 package com.sigmundgranaas.forgero.core.tool.toolpart.factory;
 
 import com.sigmundgranaas.forgero.core.material.material.PrimaryMaterial;
-import com.sigmundgranaas.forgero.core.tool.toolpart.Binding;
-import com.sigmundgranaas.forgero.core.tool.toolpart.ToolPartBinding;
+import com.sigmundgranaas.forgero.core.tool.toolpart.binding.Binding;
+import com.sigmundgranaas.forgero.core.tool.toolpart.binding.ToolPartBinding;
 
 public class ToolPartBindingBuilder extends ToolPartBuilder {
     public ToolPartBindingBuilder(PrimaryMaterial primary) {
