+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/handle/ToolPartHandle.java
@@ -1,0 +1,6 @@
+package com.sigmundgranaas.forgero.core.tool.toolpart.handle;
+
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPart;
+
+public interface ToolPartHandle extends ForgeroToolPart {
+}
