+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/head/PickaxeHead.java
@@ -1,7 +1,8 @@
-package com.sigmundgranaas.forgero.core.tool.toolpart;
+package com.sigmundgranaas.forgero.core.tool.toolpart.head;
 
 import com.sigmundgranaas.forgero.core.material.material.PrimaryMaterial;
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPartTypes;
 import com.sigmundgranaas.forgero.core.tool.toolpart.factory.ToolPartHeadBuilder;
 
 import java.util.Locale;
@@ -14,7 +15,7 @@ public class PickaxeHead extends AbstractToolPartHead {
     public PickaxeHead(ToolPartHeadBuilder toolPartHeadBuilder) {
         super(toolPartHeadBuilder.getPrimary(), toolPartHeadBuilder.getSecondary(), toolPartHeadBuilder.getHead());
     }
-    
+
     @Override
     public int getDurability() {
         return getPrimaryMaterial().getDurability();
@@ -41,8 +42,5 @@ public class PickaxeHead extends AbstractToolPartHead {
         return ForgeroToolPartTypes.HEAD;
     }
 
-    @Override
-    public int getSharpness() {
-        return 1;
-    }
+
 }
