+++ src/main/java/com/sigmundgranaas/forgero/core/tool/toolpart/head/ToolPartHead.java
@@ -1,9 +1,9 @@
-package com.sigmundgranaas.forgero.core.tool.toolpart;
+package com.sigmundgranaas.forgero.core.tool.toolpart.head;
 
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPart;
 
 public interface ToolPartHead extends ForgeroToolPart {
-    int getSharpness();
 
     int getMiningLevel();
