+++ src/main/java/com/sigmundgranaas/forgero/item/implementation/NBTFactoryImpl.java
@@ -9,9 +9,13 @@ import com.sigmundgranaas.forgero.core.tool.ForgeroTool;
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolTypes;
 import com.sigmundgranaas.forgero.core.tool.ForgeroToolWithBinding;
 import com.sigmundgranaas.forgero.core.tool.factory.ForgeroToolFactory;
-import com.sigmundgranaas.forgero.core.tool.toolpart.*;
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPart;
+import com.sigmundgranaas.forgero.core.tool.toolpart.ForgeroToolPartTypes;
+import com.sigmundgranaas.forgero.core.tool.toolpart.binding.ToolPartBinding;
 import com.sigmundgranaas.forgero.core.tool.toolpart.factory.ForgeroToolPartFactory;
 import com.sigmundgranaas.forgero.core.tool.toolpart.factory.ToolPartBuilder;
+import com.sigmundgranaas.forgero.core.tool.toolpart.handle.ToolPartHandle;
+import com.sigmundgranaas.forgero.core.tool.toolpart.head.ToolPartHead;
 import com.sigmundgranaas.forgero.item.ForgeroToolItem;
 import com.sigmundgranaas.forgero.item.NBTFactory;
 import com.sigmundgranaas.forgero.item.ToolPartItem;
